requirements.txt
PySide6
# PyQt6
# PyQt6-Qt6
PyQt6_sip
matplotlib
numpy
PyOpenGL

# contourpy
# cycler
# fonttools
# importlib_resources
# kiwisolver
# packaging
# pillow
# pyparsing
# python-dateutil
# six
# zipp
